[EGD-6319] Refactor cellular modem directory
- Removed not used files from old implementations.
- Renamed some classes to follow naming convention.
- Renamed methods to use consistent styling.
- Restructured code.

+++ module-bsp/board/linux/cellular/linux_cellular.cpp
@@ -156,8 +156,8 @@ namespace bsp
         auto timeoutTicks = pdMS_TO_TICKS(timeoutMs.count());
 
         uint32_t currentTime   = cpp_freertos::Ticks::GetTicks();
-        uint32_t timeoutNeeded = currentTime + timeoutTicks;
-        uint32_t timeElapsed   = currentTime;
+        uint64_t timeoutNeeded = currentTime + timeoutTicks;
+        uint64_t timeElapsed   = currentTime;
 
         for (;;) {
             if (timeElapsed >= timeoutNeeded) {
